Prep to announce scouting
Change old national team form form to simply direct at scouting program announcement. Add scouting link to navbar.
